# packages/contracts/openapi/dabboba.openapi.yaml
openapi: 3.1.0
info:
  title: DABBOBA API
  version: 0.1.0
  description: Shared user, commerce, and privileged operations contract. Fixture-only prototype behavior is not represented as production persistence.
servers:
  - url: http://127.0.0.1:8788
    description: Isolated local API
...
paths:
...
  /v1/kuji/rooms/{productId}/entries:
    post:
      tags: [Kuji, Commerce]
      operationId: joinKujiRoom
      description: Atomically reuses the viewer's active entry or joins the product FIFO. An empty room grants one non-renewing 180-second checkout lease.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: Existing active entry returned without creating a duplicate.
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
            X-Idempotent-Replay:
              schema: { type: string, const: "true" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KujiRoomSnapshot" }
        "201":
          description: Entry created as CHECKOUT_PENDING when free or WAITING when occupied.
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KujiRoomSnapshot" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/kuji/rooms/{productId}/entries/{entryId}:
    get:
      tags: [Kuji, Commerce]
      operationId: getOwnKujiRoomEntry
      description: Returns an uncached authoritative snapshot and atomically expires stale occupancy before promoting the next FIFO entry.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ProductId"
        - name: entryId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Current room snapshot for the entry owner.
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KujiRoomSnapshot" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Kuji, Commerce]
      operationId: leaveOwnKujiRoomEntry
      description: Idempotently cancels an owned WAITING or CHECKOUT_PENDING entry, atomically cancels any linked unpaid order and releases its stock, points, and coupon, then promotes the next FIFO entry. DRAWING or payment-authorized/paid entries return conflict.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ProductId"
        - name: entryId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Updated room snapshot with the viewer's terminal state.
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KujiRoomSnapshot" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
  parameters:
...
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
        maxLength: 120
        pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
...
  responses:
...
    Unauthorized:
      description: Authentication is required or the session is invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    NotFound:
      description: The visible resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: The requested state transition conflicts with committed state.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: { type: object, additionalProperties: true }
...
    KujiRoomEntryState:
      type: string
      enum: [WAITING, CHECKOUT_PENDING, DRAWING, COMPLETED, CANCELLED, EXPIRED]
    KujiRoomViewer:
      type: object
      additionalProperties: false
      required: [entryId, state, position, peopleAhead, checkoutExpiresAt]
      properties:
        entryId: { type: string, format: uuid }
        state: { $ref: "#/components/schemas/KujiRoomEntryState" }
        position: { type: [integer, "null"], minimum: 1 }
        peopleAhead: { type: integer, minimum: 0 }
        checkoutExpiresAt: { type: [string, "null"], format: date-time }
    KujiRoomActiveEntry:
      type: object
      additionalProperties: false
      required: [displayName, phase, checkoutExpiresAt]
      properties:
        displayName: { type: string, minLength: 2, maxLength: 40 }
        phase: { type: string, enum: [CHECKOUT_PENDING, DRAWING] }
        checkoutExpiresAt: { type: [string, "null"], format: date-time }
    KujiRoomWaitingPerson:
      type: object
      additionalProperties: false
      required: [entryId, displayName, position, isViewer]
      properties:
        entryId: { type: string, format: uuid }
        displayName: { type: string, minLength: 2, maxLength: 40 }
        position: { type: integer, minimum: 1 }
        isViewer: { type: boolean }
    KujiRoomRecentActivity:
      type: object
      additionalProperties: false
      required: [id, displayName, prizeName, prizeImageUrl, rarity, committedAt]
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string, minLength: 2, maxLength: 40 }
        prizeName: { type: string, minLength: 1, maxLength: 240 }
        prizeImageUrl: { type: [string, "null"], format: uri }
        rarity: { type: string, minLength: 1, maxLength: 40 }
        committedAt: { type: string, format: date-time }
    KujiRoomSnapshot:
      type: object
      additionalProperties: false
      required: [serverNow, version, productId, viewer, active, waitingCount, waitingPeople, recentActivity]
      properties:
        serverNow: { type: string, format: date-time }
        version: { type: integer, minimum: 1 }
        productId: { type: string, maxLength: 120 }
        viewer: { $ref: "#/components/schemas/KujiRoomViewer" }
        active:
          oneOf:
            - $ref: "#/components/schemas/KujiRoomActiveEntry"
            - type: "null"
        waitingCount: { type: integer, minimum: 0 }
        waitingPeople:
          type: array
          items: { $ref: "#/components/schemas/KujiRoomWaitingPerson" }
        recentActivity:
          type: array
          maxItems: 8
          items: { $ref: "#/components/schemas/KujiRoomRecentActivity" }
...